+++ packages/playground/resolve/vite.config.js
@@ -2,7 +2,7 @@ const virtualFile = '@virtual-file'
 const virtualId = '\0' + virtualFile
 
 const customVirtualFile = '@custom-virtual-file'
-const { a } = require('./config-dep');
+const { a } = require('./config-dep')
 
 module.exports = {
   resolve: {