+++ packages/vite/src/node/plugins/ssrRequireHook.ts
@@ -1,6 +1,6 @@
 import MagicString from 'magic-string'
-import { ResolvedConfig } from '..'
-import { Plugin } from '../plugin'
+import type { ResolvedConfig } from '..'
+import type { Plugin } from '../plugin'
 import { arraify } from '../utils'
 
 /**
@@ -12,6 +12,7 @@ export function ssrRequireHookPlugin(config: ResolvedConfig): Plugin | null {
   if (
     config.command !== 'build' ||
     !config.resolve.dedupe?.length ||
+    config.ssr?.noExternal === true ||
     isBuildOutputEsm(config)
   ) {
     return null