+++ packages/vite/src/node/server/middlewares/indexHtml.ts
@@ -1,18 +1,19 @@
 import fs from 'fs'
 import path from 'path'
 import MagicString from 'magic-string'
-import { AttributeNode, NodeTypes } from '@vue/compiler-dom'
-import { Connect } from 'types/connect'
+import type { AttributeNode } from '@vue/compiler-dom'
+import { NodeTypes } from '@vue/compiler-dom'
+import type { Connect } from 'types/connect'
+import type { IndexHtmlTransformHook } from '../../plugins/html'
 import {
   addToHTMLProxyCache,
   applyHtmlTransforms,
   assetAttrsConfig,
   getScriptInfo,
-  IndexHtmlTransformHook,
   resolveHtmlTransforms,
   traverseHtml
 } from '../../plugins/html'
-import { ResolvedConfig, ViteDevServer } from '../..'
+import type { ResolvedConfig, ViteDevServer } from '../..'
 import { send } from '../send'
 import { CLIENT_PUBLIC_PATH, FS_PREFIX } from '../../constants'
 import { cleanUrl, fsPathFromId, normalizePath, injectQuery } from '../../utils'
@@ -121,12 +122,20 @@ const devHtmlHook: IndexHtmlTransformHook = async (
         addToHTMLProxyCache(config, url, scriptModuleIndex, contents)
 
         // inline js module. convert to src="proxy"
+        const modulePath = `${
+          config.base + htmlPath.slice(1)
+        }?html-proxy&index=${scriptModuleIndex}.js`
+
+        // invalidate the module so the newly cached contents will be served
+        const module = server?.moduleGraph.getModuleById(modulePath)
+        if (module) {
+          server?.moduleGraph.invalidateModule(module)
+        }
+
         s.overwrite(
           node.loc.start.offset,
           node.loc.end.offset,
-          `<script type="module" src="${
-            config.base + url.slice(1)
-          }?html-proxy&index=${scriptModuleIndex}.js"></script>`
+          `<script type="module" src="${modulePath}"></script>`
         )
       }
     }